small patch who wants their filenames to be randomized.
in config file "filenameRandomize" variable added. set false as default.

in core/class/browser.php under act_upload function added if clause to check if config["filenameRandomize"] set to true. if so call getRandomHash to get random string and set it as filename.

for extensiton pathinfo would be more secure to get extension of file.